+++ e_ood/__init__.py
@@ -3,6 +3,6 @@ __version__ = '0.1.40a1'
 # !__version__ must be set on the second line, for setup.py!
 
 from .analyze import Analyzer  # noqa
-from .db import ReportedUpdateTypes, VersionDB  # noqa
+from .db import ReportedUpdateTypes, PackageVersionClassifications  # noqa
 from .pypi import AvailablePackageVersions  # noqa
 from .virtualenv import EnvPackages  # noqa